+++ misc/update_DESCRIPTION_NEWS.R
@@ -290,13 +290,18 @@ my_news$add_bullet(c("removing ASCII characters"))
 my_desc$bump_version("major")
 my_news$add_version(my_desc$get_version())
 my_news$add_subtitle("update function")
-my_news$add_bullet(c("removing dive and burglr"))
+my_news$add_bullet(c("removing dive and burglr",
+                     "adding new function 'save_rds_archive'"))
 
 my_news$add_subtitle("internal changes")
 my_news$add_bullet(c("including github actions",
                      "removing other CI"))
 my_news$add_subtitle("styling")
 my_news$add_bullet(c("fixing lints"))
+
+my_desc$add_author('Lukas', 'Feick',
+                   email = 'lukas.feick@statworx.com',
+                   role = c('aut'))
 
 # save everything ---------------------------------------------------------